requirements.txt
djangorestframework==3.15.2
drf-spectacular==0.27.2
mysqlclient==2.2.4